revised week_of_the_year()

+++ chrono.h
@@ -29,7 +29,7 @@ namespace Chrono {
 		long int days_since_jan_1_1970;
 	};
 	bool leapyear(int yy);
-	double week_of_the_year(const Date& d);
+	int week_of_the_year(const Date& d);
 	bool is_date(int yy, Month m, int d);
 	Chrono::Day day_of_the_week(const Date& d);
 	Date next_workday(const Date& d);
